SmileDemo/requirements.txt
streamlit
streamlit-webrtc
fastapi
uvicorn[standard]
python-multipart
dlib
face_recognition
numpy
Pillow
requests
mediapipe
tensorflow